Starting work on Emulator
Looking up the specs of the platform on google I found that the CPU
has 16 8-bit registers. a 64 byte stack. Among other things...

Accodrind the spec the frame buffer is has a bit-depth of 1 so I'm
going to write an extremely compact framebuffer that can fit in cache
of any modern CPU snugly.